@@ -12,25 +12,25 @@
   },
   "devDependencies": {
     "autoprefixer": "^6.7.7",
-    "babel-cli": "^6.24.0",
-    "babel-core": "^6.24.0",
+    "babel-cli": "^6.24.1",
+    "babel-core": "^6.24.1",
     "babel-loader": "^6.4.1",
     "babel-plugin-transform-object-rest-spread": "^6.23.0",
     "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-es2015": "^6.24.0",
-    "babel-preset-react": "^6.23.0",
-    "css-loader": "^0.27.1",
-    "eslint": "^3.18.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "css-loader": "^0.28.0",
+    "eslint": "^3.19.0",
     "eslint-config-airbnb": "^14.1.0",
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-jsx-a11y": "^4.0.0",
-    "eslint-plugin-react": "^6.10.0",
+    "eslint-plugin-react": "^6.10.3",
     "extract-text-webpack-plugin": "^2.1.0",
-    "file-loader": "^0.10.1",
+    "file-loader": "^0.11.1",
     "html-loader": "^0.4.5",
     "jest": "^19.0.2",
     "loader-utils": "^1.1.0",
-    "node-sass": "^4.5.0",
+    "node-sass": "^4.5.2",
     "nodemon": "^1.9.2",
     "null-loader": "^0.1.0",
     "postcss-loader": "^1.3.3",
@@ -39,9 +39,9 @@
     "rimraf": "^2.6.1",
     "sass-loader": "^6.0.3",
     "source-map-support": "^0.4.14",
-    "style-loader": "^0.14.1",
+    "style-loader": "^0.16.1",
     "url-loader": "^0.5.8",
-    "webpack": "^2.2.1",
+    "webpack": "^2.3.3",
     "webpack-dev-server": "^2.4.2"
   },
   "dependencies": {
@@ -53,12 +53,12 @@
     "ejs": "^2.5.6",
     "express": "^4.15.2",
     "raf": "^3.1.0",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "router1": "^0.6.0",
-    "router1-react": "^0.7.0",
-    "rx-react-container": "^0.2.0",
-    "rxjs": "^5.2.0"
+    "react": "^15.5.3",
+    "react-dom": "^15.5.3",
+    "router1": "^0.6.1",
+    "router1-react": "^0.7.1",
+    "rx-react-container": "^0.2.1",
+    "rxjs": "^5.3.0"
   },
   "jest": {
     "roots": [
