@@ -1,4 +1,5 @@
-import { Observable } from 'rxjs';
+import { fromPromise } from 'rxjs/observable/fromPromise';
+import { of } from 'rxjs/observable/of';
 import { isPromise } from 'rxjs/util/isPromise';
 import { isFunction } from 'rxjs/util/isFunction';
 
@@ -6,6 +7,6 @@ const isObservable = v => isFunction(v.subscribe);
 
 export default function (data) {
   if (isObservable(data)) return data;
-  if (isPromise(data)) return Observable.fromPromise(data);
-  return Observable.of(data);
+  if (isPromise(data)) return fromPromise(data);
+  return of(data);
 }
