@@ -1,10 +1,10 @@
-import { Observable } from 'rxjs';
+import { of} from 'rxjs/observable/of';
 import toObservable from './toObservable';
 
 describe('toObservable', () => {
   it('wraps object', done => {
     toObservable({ a: 1 })
-      .forEach(v => {
+      .subscribe(v => {
         expect(v).toEqual({ a: 1 });
         done();
       });
@@ -12,15 +12,15 @@ describe('toObservable', () => {
 
   it('wraps promise', done => {
     toObservable(new Promise(r => r({ a: 1 })))
-      .forEach(v => {
+      .subscribe(v => {
         expect(v).toEqual({ a: 1 });
         done();
       });
   });
 
   it('wraps observable', done => {
-    toObservable(Observable.of({ a: 1 }))
-      .forEach(v => {
+    toObservable(of({ a: 1 }))
+      .subscribe(v => {
         expect(v).toEqual({ a: 1 });
         done();
       });
